requirements.txt
jupyterlab
numpy==1.23.1
opencv-python
Pillow
torch
torchvision
